requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.104.0
uvicorn>=0.24.0
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0
pillow>=8.0.0
trimesh>=3.20.0
scikit-image>=0.19.0
pydantic>=2.0.0
matplotlib>=3.5.0
pyyaml>=5.4.0
requests>=2.25.0